refactor: rename fold
